[feature/PLAYER-3147] remove redundant list item components, re-factor MultiAudioTab, add tests
